Move AccessToken to rxnakadi package.
This is far more secure when using together with undertaking. If the classes' ABI diverge, there will be nasty runtime errors.
